chore: reformat, replace tabs with 4-space indent
This changes the clang-format configuration to use 4-space
indent and runs it against all .cc files.

This is a purely mechanical source translation, and includes
no behavioral or other functional changes.

I'm not happy with how, e.g., the dispatch tables were
formatted, but we can play around with that and in the
meantime, I think it is more important to have consistent
automation around this.

Fixes #13

Signed-off-by: Dan Cross <cross@m-net.arbornet.org>