tests: fix `str_test` and add a `test` make target
`str_test.cc` was failing due to build ambiguity in the
`Contains` test case (`str::contains` is overloaded for
both vectors of string and string_view).

Add a `test` target to `GNUmakefile` to run tests.  As
we add new test cases, we merely need to add these to
the `TESTS` variable and they should be automatically
run; this facilitates CI.

Signed-off-by: Dan Cross <cross@m-net.arbornet.org>